Extracting re-usable code into `sov-soak-lib`
