+++ scripts/switcheroo/presets/examples.json
@@ -2,5 +2,6 @@
 	"examples/demo-rollup",
 	"examples/demo-rollup/rest-api-load-testing",
 	"examples/demo-rollup/sov-benchmarks",
-	"examples/demo-simple-stf"
+	"examples/demo-simple-stf",
+	"crates/utils/sov-soak-lib"
 ]
